+++ NEWS.md
@@ -12,6 +12,7 @@ CREATE INDEX idx_medreq_requester_barcode ON ${database.defaultSchemaName}.media
 * [FOLSPRINGS-200](https://folio-org.atlassian.net/browse/FOLSPRINGS-200) commons-fileupload 1.6.0 fixing DoS from unlimited multipart headers CVE-2025-48976
 * [FOLSPRINGS-202](https://folio-org.atlassian.net/browse/FOLSPRINGS-202) change warning log to debug for execution scope not set up message
 * [FOLSPRINGS-208](https://folio-org.atlassian.net/browse/FOLSPRINGS-208) Bump Rhino from 1.8.0 to 1.8.1 fixing CVE-2025-66453
+* [FOLSPRINGS-209](https://folio-org.atlassian.net/browse/FOLSPRINGS-209) Reduce logging during context switching
 
 ### cql submodule
 * [FOLSPRINGS-185](https://folio-org.atlassian.net/browse/FOLSPRINGS-185) Implement case insensitive accents ignoring CQL queries
@@ -23,6 +24,10 @@ CREATE INDEX idx_medreq_requester_barcode ON ${database.defaultSchemaName}.media
 * [FOLSPRINGS-195](https://folio-org.atlassian.net/browse/FOLSPRINGS-195) Add headers parameter to executeSystemUserScoped method
 * [FOLSPRINGS-201](https://folio-org.atlassian.net/browse/FOLSPRINGS-195) Deprecate folio-spring-system-user submodule
 * [FOLSPRINGS-207](https://folio-org.atlassian.net/browse/FOLSPRINGS-207) Add executeSystemUserScoped method with userId
+* [MODELINKS-367](https://folio-org.atlassian.net/browse/MODELINKS-367) Add executeAsyncSystemUserScoped method with userId
+
+### folio-spring-tenant-settings
+* [FOLSPRINGS-210](https://folio-org.atlassian.net/browse/FOLSPRINGS-210) Add tenant-settings submodule for managing tenant-specific configuration
 
 ## 9.0.0 2025-02-28
 * [FOLSPRINGS-188](https://folio-org.atlassian.net/browse/FOLSPRINGS-188) Upgrade to Java 21